enable SegWit trigger, add initial TRIG regression test (MVHF-CORE-DES-TRIG-6)
Triggering is extending to include SegWit activation

A qa test (mvf-core-trig.py) is added which so far covers only the activation during generation of a chain. It currently takes about 10 sec to execute on my machine.

The test still needs to be extended to cover loading of the chain from disk at startup (again with various cases of block height and SegWit activation/non-activation).

De-activation is also not yet tested (although it falls under TRIG). The test may be split up in future to test these separate functionalities.