add basic persistence using btcfork.conf file (MVHF-CORE-SW-REQ-2-5 / MVHF-CORE-DES-TRIG-10)
The settings stored in the config file are not read back in during program 
initialization. This might be necessary later on, to determine whether the
forkport setting (to be added) matches the port setting or not.

A mismatch would be an indication to prefer the forkport setting in order to
connect to the forked network.

Whether or not the fork has been activated in a prior run is stored in the global
variable 'wasMVFHardForkPreviouslyActivated'.

The mvf-core-trig.py and walletbackupauto.py tests are extended:
- mvf-core-trig.py now tests whether the fork file is written and the detection works
- the wallet backup tests remove the btcfork.conf prior to node restarts so that they can test the .old file writing correctly

In future, the application should probably output a message at startup/shutdown and
in the GUI if it detects this file, to prompt the user to integrate the relevant settings into
the main configuration file (bitcoin.conf).