fix possible overflow in calculation of new difficulty target
This occurred on regtest where the POW limit results in very large target,
which if multiplied by the time in seconds of even a reasonably small lookback
window can overflow the 256 bit integer and result in an unexpected (and
often large) adjustment in the opposite direction.

This resulted in difficulty becoming too large for the test framework to get
a response in time, which gives a 'timed out' error.

Due to slightly different rounding, one existing test in pow_tests.cpp had
to be corrected in the least significant decimal.