add tool under contrib/ which allows parallelizing C++ unit tests
[this is a merge from MVF-BU commit c757a8299e9073910ad5e09e94d14d55a89ee865]

contrib/testtools/gtest-parallel-bitcoin is a rough adaptation of Google's
test sharding wrapper at https://github.com/google/gtest-parallel .

It has been adapted to use the Boost Test framework's '--list_content'
command line parameter, and other Google Test dependencies have been
removed.

It can be run by simply giving it the path to the test_bitcoin executable,
which it then queries to obtain a list of runnable tests.
This only works if Bitcoin has been compiled against Boost 1.59 or later.
On earlier versions, the executable would unfortunately ignore the '--list_content'
option and start running all the tests it knows.

Running with the '--format=raw' option provides more detailed test output.

The parallelization wrapper creates and maintains a score database of test 
execution times in ~/.gtest-parallel-bitcoin-times .
When it executes tests in repeated runs, it does so in order from longest to
shortest test.

Testing on a desktop machine with sufficient cores shows that the total test
execution time when using this parallelization script is is reduced to
approximately the running time of the slowest unit test.