more work on option / script parameter handling, including timekeeping and tabulation at end
some things don't work right yet, e.g. you cannot specify `sometest --testspecificargument` and then have the runner execute the right version - it will execute the first one it finds that matches sometest (although with the arguments passed on), but it will not execute later entries of the same test script anymore.

A better design for the entire runner script is needed, also w.r.t. argument handling. If test names are specified in as options, then all subsequent parameters should be bound to them unless they are known to be parameters specific to the the runner.
Also, the whole split-up of help between the runner and tests (which can have specific args) is bonkers when it comes to printing out usage.
The runner should not call a test to print out general options.

Core has done some good work on version in 0.13 that is capable of executing tests in parallel - a rework of the runner should take some of those improvements onboard.