add improvements aimed at facilitating Travis and testing of disabled/skipped tests
this mirrors MVF-BU PR : https://github.com/BTCfork/hardfork_prototype_1_mvf-bu/pull/49

the long mvf-core-retarget.py has been moved to the extended section, and it now accepts a --quick argument which runs only a subset of the post-fork retargeting. This brings the test's duration down significantly, hopefully enough to allow us to run Travis with it.

Also added better handling of rpc-tests.py options -- now it is possible to specify a test plus its option on the command line and it will be run correctly (including running the test only with that option, not twice if included in both regular and extended).

A new -force-enable (-f) option is added to execute Disabled/Skipped tests, which is useful when you don't want to permanently re-enable them, but tentatively test them.