initial merge of btcfork.conf parsing (PR58) from MVF-BU
This implements reading of the values in the btcfork.conf file on startup if it is present.

The idea here was that when the fork happens, this file is written out with the key parameters used to fork. This helps the client to recognize at startup that it has forked previously, and can then initialize itself correctly.

Ultimately, the user will be requested to transfer some of the final settings to his bitcoin.conf file, and can then remove the btcfork.conf file entirely. This part is not yet implemented.