initial merge of master into feature/csig_experimentation
util.py was changed to extract program constants from mvf-core-globals.h instead of mvf-core.h (this is due to the factoring out of global vars into mvf-core-globals.* in this PR).

The rest of the merge were just routine adaptations to this factoring.

Both unit tests and regular qa tests passed in local Linux64 dev env.
Extended qa tests still need to be completed.